tests/wpt/web-platform-tests/portals/portals-adopt-predecessor.html

<!DOCTYPE html>
<title>Tests that a portal can adopt its predecessor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(function(t) {
    var test = "adopt-once";
    var bc = new BroadcastChannel(`test-${test}`);
    bc.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, "passed");
      bc.close();
    });
    window.open(`resources/portals-adopt-predecessor.html?test=${test}`);
  }, "Tests that a portal can adopt its predecessor.");

  async_test(function(t) {
    var test = "adopt-twice";
    var bc = new BroadcastChannel(`test-${test}`);
    bc.onmessage = t.step_func_done(function(e) {
      assert_equals(e.data, "passed");
      bc.close();
    });
    window.open(`resources/portals-adopt-predecessor.html?test=${test}`);
  }, "Tests that trying to adopt the predecessor twice will throw an exception.");
</script>
